finished RecordList refactor
The old design of RecordList would attempt to swallow as much weird data as
possible. The new design throws parsing exceptions when it find something it
doesn't expect. These can be ignored if needed, but it's probably best to
examine the error and figure out the proper behaviour.

h5_generate_cache is included in this commit because it's closely tied to
RecordList and exception handling has been changed slightly.