added FileHandler.GetEventValsByFID method
This makes it possible to query a FileHandler for specific FIDs. This is useful
because the FIDs in a Gather may not be the same as the ones in a pickfile. By
using this method rather than the old ones, the traces can be matched up
correctly.

Also modified Gather.LoadPicks to use the new method.