separate picking attributes and methods
picking attributes and methods have now been placed in the PickableGather class,
so that they are cleanly seperated from the CommonMidpointGather and
CommonOffsetGather. This removes cruft from code that doesn't deal with picking,
and makes it simpler to experiment with alternative picking algorithms. This
will temporarily break icepick.py, which now needs to instantiate
PickableCOGather objects.