Simple hack to fix issue #6
It seems that `self.fig.canvas.toolbar.destroy()` isn't safe for all possible
backends. A work-around is to dynamically set the rcParams, although this could
cause the effect to spread to other windows if used in a script with other
matplotlib calls.

It seems reasonable that `irlib.app` will be used stand-alone, so I did it for
now with the rcParams adjustment happening on load of the
`irlib.app.components.module`.