exchange `flip_lons` kwarg -> `eastern_hemisphere`
more self-explanatory

not that the meaning is opposite what it was before

Background
----------

BSI IceRadar does not currently store the sign of the longitude. The metadata
tag is `Long_ W`, so it is presumably meant to be western hemisphere by default,
and the absence of a minus in the eastern hemisphere is a bug.